Intro to Computer Science/Week 10 - The End/project/requirements.txt
Flask
Flask-Session
pytz
requests
datetime
Authlib
python-dotenv
google-api-python-client
google-auth